Add files via upload
Here's a toy backprop file used to better understand how LLMs work. It's designed to do a few things inside a single file: tokenizer, toy training text, creation of arrays with weights (including outputting files that can be re-used in a future instance to keep training), prediction of next token. The only dependency is NumPy. Again, this is a toy example but powerful that the entire process can be distilled, for learning purposes, into a single file that can be run to completion including all aspects of the process and not requiring any additional data or files (other than NumPy) to run. There is also something quite rewarding in knowing that these LLM training scripts do not need to be run on H100 nodes and that there is a process, albeit slow, whereby weights can be stored in a non-compressed way (i.e. human readable) and that the tokenization can be made in an arbitrarily simple way that still functions. Obviously this model will not actually be useful for prediction, but this was inspired by Andrej Karpathy's recent (February 2025) YouTube video. Code generation is 100% Grok2 for full disclosure and took about 2 hours with breaks to finish and approximately 25 prompts.